# openapi.yaml
openapi: 3.0.3
info:
  title: gdl-api
  description: gdl-api express api server + frontend + bots
  version: 0.2.31
  license:
    name: MIT
servers:
  - url: https://gdl.mcalec.dev/api
    description: production server
  - url: http://localhost:3030/api
    description: local server
...
paths:
...
  /user/announcements:
    get:
      tags:
        - User
      summary: Retrieve public announcements
      description: Get a list of announcements for all users, sorted by creation date
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    title:
                      type: string
                    message:
                      type: string
                    severity:
                      type: string
                      enum: [info, warning, error]
                    author:
                      type: string
                    created:
                      type: string
                      format: date-time
                    modified:
                      type: string
                      format: date-time
                    uuid:
                      type: string
        "500":
          description: Internal server error
  /admin/:
    get:
      tags:
        - Admin
      summary: Get admin dashboard statistics
      description: Retrieve comprehensive admin statistics including uptime, user counts, and storage usage
      responses:
        "200":
          description: Admin statistics
        "401":
          description: User not authenticated or not an admin
        "500":
          description: Internal server error
...
  /admin/user/{id}:
    get:
      tags:
        - Admin
      summary: Get user information
      description: Retrieve detailed information about a specific user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID
      responses:
        "200":
          description: User information
    put:
      tags:
        - Admin
      summary: Update user information
      description: Modify user account details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID
      responses:
        "204":
          description: User updated successfully
  /admin/user/{id}/action:
    post:
      tags:
        - Admin
      summary: Perform user actions (ban, mute, suspend)
      description: Execute administrative actions on user accounts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [ban, mute, suspend]
                ip:
                  type: string
                  description: Optional IP address to ban
      responses:
        "201":
          description: Action completed successfully
        "401":
          description: User not authorized
...
  /admin/tags/{id}:
    post:
      tags:
        - Admin
      summary: Create a new tag
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Tag created successfully
    put:
      tags:
        - Admin
      summary: Update a tag
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Tag updated successfully
    delete:
      tags:
        - Admin
      summary: Delete a tag
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Tag deleted successfully
...
  /admin/dmca/{id}:
    put:
      tags:
        - Admin
      summary: Update a DMCA/takedown request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Takedown request updated
...
  /admin/announcements:
    get:
      tags:
        - Admin
      summary: Get all announcements
      description: Retrieve all announcements for administrators
      responses:
        "200":
          description: List of announcements
    post:
      tags:
        - Admin
      summary: Create a new announcement
      description: Create a new system-wide announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                message:
                  type: string
                severity:
                  type: string
                  enum: [info, warning, error]
      responses:
        "201":
          description: Announcement created successfully
  /admin/announcements/{uuid}:
    put:
      tags:
        - Admin
      summary: Update an announcement
      description: Update the title, message, or severity of an existing announcement
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
          description: Announcement UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                message:
                  type: string
                severity:
                  type: string
                  enum: [info, warning, error]
      responses:
        "204":
          description: Announcement updated successfully
        "400":
          description: Invalid UUID or missing required fields
        "404":
          description: Announcement not found
    delete:
      tags:
        - Admin
      summary: Delete an announcement
      description: Remove an announcement from the system
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
          description: Announcement UUID
      responses:
        "204":
          description: Announcement deleted successfully
        "400":
          description: Invalid UUID
        "401":
          description: User not authorized as admin
        "404":
          description: Announcement not found
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authenticated requests
security:
  - cookieAuth: []